scripts/requirements.txt
openai
requests
python-dotenv
google-generativeai
fal-client
